Fix long-standing bug whereby popups appear during drawing
This was a difficult bug to fix, but the solution seems robust.
The fix works by disabling the interactivity of each layer that has popups by using map.off as soon as drawing is started, and reinstating using map.on when drawing stops or is stopped.
See: https://docs.mapbox.com/mapbox-gl-js/api/map/#map#off
To facilitate this, the popup handlers have to be registered to a global registry, so that map.off and map.on have a handle (see third argument) that can be referenced; see: https://stackoverflow.com/a/45665068/180733
This has to be hooked into the mapbox-gl-draw events as documented at: https://github.com/mapbox/mapbox-gl-draw/blob/main/docs/API.md#events
If adding new buttons using the mapbox-gl-draw events, they will also need to set the _drawingHappening state as relevant.

A variety of other strategies were tried, including the following, but none worked:
- Attempting to avoid the click propagating, but e.stopPropagation does not exist in this context so this generates an error; see mapbox-gl-js issue #9369
- Stopping the popup opening at the point of querying for features (using queryRenderedFeatures) at the clicked location; see: https://stackoverflow.com/questions/47576527/
- Closing the popup immediately on opening (which would probably have given flicker)
- Overriding the map click somehow; see: https://stackoverflow.com/questions/45298202/ , which though helpfully pointed to the use of map.off/map.on