On style (background map) change, reload layers
This was a very difficult reworking.

Previously, a basemap change was set to trigger a data reload.
However, at some point, it seems that Mapbox GL JS causes the source to be lost, not just the layer.
This means a setData call fails, because there is no source to load to.

This commit forces a full layer reload at the point of basemap change.
However, that itself is buggy, because removeLayer seems to be slow to complete.
As a result, adding a layer fails in odd ways.
So, the new layer addition is now placed in a timeout.

The timings may need some adjustment, as slower basemaps like OS outdoor may not always load in time.

This is all rather unsatisfactory, but it mostly works now.
The root of these problems is due to timing and other bugs as outlined in:
https://github.com/mapbox/mapbox-gl-js/issues/8691